Update TIP to v2.12.2 (#45)
### 2.12.2

- Add automatic handling of unnecessary image download errors when the download has finished loading
  - It is not uncommon for a service/CDN to yield an error after the final byte of the response has been loaded by the client
  - The consequence of treating a successful load as a failure is that upon next fetch for that image an unnessecary network request will trigger:
    - For image resumption supported loads, resumes the load beyond the available byte range
    - For full image loads, a redundant download
  - When TIP sees an error on image download but has all the bytes (looking at `Content-Length` header), TIP now posts `TIPProblemImageDownloadedWithUnnecessaryError` problem

Also, many misc bug fixes

+++ TwitterImagePipeline/TIPImageFetchMetrics.m
@@ -242,7 +242,7 @@ static void _hit(SELF_ARG,
 {
     const size_t count = sizeof(_infos) / sizeof(_infos[0]);
     NSMutableArray *array = [[NSMutableArray alloc] initWithCapacity:count];
-    for (TIPImageLoadSource i = 1; i <= count; i++) {
+    for (TIPImageLoadSource i = 1; (size_t)i <= count; i++) {
         TIPImageFetchMetricInfo *info = _infos[i - 1];
         if (info) {
             [array addObject:info];
